+++ src/services/axios/sensitivityCalculator/calculateSensitivityCalculatorResults.ts
@@ -35,7 +35,7 @@ export default function calculateSensitivityCalculatorResults(
 
   const weightedSensitivity = isLow()
     ? getWeightedSensitivityLOW(response, isZoom())
-    : getWeightedSensitivityMID(observation, response, isZoom());
+    : getWeightedSensitivityMid(response, isZoom(), observation);
 
   const confusionNoise = getConfusionNoise(response, isZoom());
 
@@ -76,21 +76,20 @@ export default function calculateSensitivityCalculatorResults(
   const confusionNoiseDisplay = sensCalHelpers.format.convertReturnedSensitivityToDisplayValue(
     confusionNoise
   );
-  const weightedSensitivityDisplay = isSensitivitySupplied
+  const weightedSensitivityDisplay = isSensitivitySupplied()
     ? convertSuppliedSensitivityToDisplayValue(weightedSensitivity)
     : sensCalHelpers.format.convertReturnedSensitivityToDisplayValue(weightedSensitivity);
   const totalSensitivityDisplay = sensCalHelpers.format.convertReturnedSensitivityToDisplayValue(
-    totalSensitivity
+    isZoom() ? spectralTotalSensitivity : totalSensitivity
   );
   const beamSizeDisplay = { value: beamSize, units: 'arcsec2' };
   const sbsDisplay = { value: sbs, units: 'k' };
 
-  const spectralConfusionNoiseDisplay =
-    observation.type === TYPE_CONTINUUM
-      ? sensCalHelpers.format.convertReturnedSensitivityToDisplayValue(spectralConfusionNoise)
-      : { value: '', unit: '' };
+  const spectralConfusionNoiseDisplay = sensCalHelpers.format.convertReturnedSensitivityToDisplayValue(
+    spectralConfusionNoise
+  );
 
-  const spectralWeightedSensitivityDisplay = isSensitivitySupplied
+  const spectralWeightedSensitivityDisplay = isSensitivitySupplied()
     ? convertSuppliedSensitivityToDisplayValue(spectralWeightedSensitivity)
     : sensCalHelpers.format.convertReturnedSensitivityToDisplayValue(spectralWeightedSensitivity);
 
@@ -112,17 +111,23 @@ export default function calculateSensitivityCalculatorResults(
     section1: [
       {
         field: `${observationTypeLabel}SensitivityWeighted`,
-        value: weightedSensitivityDisplay?.value.toString(),
+        value: isZoom()
+          ? spectralWeightedSensitivityDisplay?.value.toString()
+          : weightedSensitivityDisplay?.value.toString(),
         units: weightedSensitivityDisplay?.unit
       },
       {
         field: `${observationTypeLabel}ConfusionNoise`,
-        value: confusionNoiseDisplay?.value.toString(),
+        value: isZoom()
+          ? spectralConfusionNoiseDisplay?.value.toString()
+          : confusionNoiseDisplay?.value.toString(),
         units: confusionNoiseDisplay?.unit
       },
       {
         field: `${observationTypeLabel}TotalSensitivity`,
-        value: totalSensitivityDisplay?.value.toString(),
+        value: isZoom()
+          ? spectralTotalSensitivityDisplay?.value.toString()
+          : totalSensitivityDisplay?.value.toString(),
         units: totalSensitivityDisplay?.unit
       },
       {
@@ -235,21 +240,26 @@ const getSpectralSurfaceBrightnessLOW = (
 
 /******************************************* MID ********************************************/
 
-const getWeightedSensitivityMID = (
-  observation: Observation,
+const getWeightedSensitivityRawValueMid = (
   response: SensitivityCalculatorAPIResponseMid,
-  isZoom: boolean
+  isZoom: boolean,
+  observation: Observation
 ) => {
   const recCalc = isZoom ? response?.calculate?.data[0] : response?.calculate?.data;
-  if (recCalc?.spectral_sensitivity?.value) {
-    const sensitivity: number = recCalc?.spectral_sensitivity.value;
-    const factor: number = isZoom
-      ? response.weighting[0].weighting_factor
-      : response.weighting.weighting_factor;
-    return sensitivity * factor;
+  const recWeight = isZoom ? response?.weighting[0] : response?.weighting;
+  if (recCalc?.continuum_sensitivity?.value) {
+    return recCalc.continuum_sensitivity!.value! * recWeight.weighting_factor * 1e6;
   } else {
     return observation?.supplied.value;
   }
+};
+
+const getWeightedSensitivityMid = (
+  response: SensitivityCalculatorAPIResponseMid,
+  isZoom: boolean,
+  observation: Observation
+) => {
+  return getWeightedSensitivityRawValueMid(response, isZoom, observation);
 };
 
 const getBeamSizeMID = (response: SensitivityCalculatorAPIResponseMid, isZoom): string => {
@@ -267,21 +277,26 @@ const getBeamSizeMID = (response: SensitivityCalculatorAPIResponseMid, isZoom): 
 
 /* -------------- */
 
+const getSpectralWeightedSensitivityRawValueMid = (
+  observation: Observation,
+  response: SensitivityCalculatorAPIResponseLow,
+  isZoom: boolean
+) => {
+  const recCalc = isZoom ? response?.calculate?.data[0] : response?.calculate?.data;
+  const recWeightLine = isZoom ? response?.weighting[0] : response?.weightingLine[0];
+  if (recCalc?.spectral_sensitivity?.value) {
+    return recCalc.spectral_sensitivity?.value! * recWeightLine.weighting_factor * 1e6;
+  } else {
+    return observation?.supplied.value;
+  }
+};
+
 const getSpectralWeightedSensitivityMID = (
   observation: Observation,
   response: SensitivityCalculatorAPIResponseLow,
   isZoom: boolean
 ) => {
-  const rec = isZoom ? response.weighting[0] : response.weightingLine[0];
-  if (!rec) {
-    return 0;
-  }
-  const calc = isZoom ? response.calculate.data[0] : response.calculate.data;
-  if (calc?.spectral_sensitivity?.value) {
-    return calc.spectral_sensitivity?.value * rec.weighting_factor;
-  } else {
-    return observation?.supplied.value;
-  }
+  return getSpectralWeightedSensitivityRawValueMid(observation, response, isZoom);
 };
 
 const getSpectralBeamSizeMID = (response: SensitivityCalculatorAPIResponseMid, isZoom: boolean) => {
@@ -311,7 +326,7 @@ const getSpectralSurfaceBrightnessMID = (
 /********************************************* COMMON ***********************************************/
 
 const getSensitivity = (confusionNoise: number, weightedSensitivity: number): number =>
-  sensCalHelpers.calculate.sqrtOfSumSqs(confusionNoise * 1e6, weightedSensitivity);
+  sensCalHelpers.calculate.sqrtOfSumSqs(confusionNoise, weightedSensitivity);
 
 const getConfusionNoise = (response: SensitivityCalculatorAPIResponseLow, isZoom): number => {
   const rawConfusionNoise = getRawConfusionNoise(response, isZoom);
@@ -325,9 +340,9 @@ const getSpectralConfusionNoise = (
   isZoom: Boolean
 ) => {
   const rawSpectralConfusionNoise = getSpectralRawConfusionNoise(response, isZoom);
-  return sensCalHelpers.format.convertConfusionNoiseRawValueInuJy(
-    Number(rawSpectralConfusionNoise)
-  );
+  return rawSpectralConfusionNoise
+    ? sensCalHelpers.format.convertConfusionNoiseRawValueInuJy(Number(rawSpectralConfusionNoise))
+    : 0;
 };
 
 const getRawConfusionNoise = (response: SensitivityCalculatorAPIResponseLow, isZoom): number => {