+++ CHANGELOG.md
@@ -15,6 +15,8 @@ Future
 * STAR-640:
    - Display only most recent proposal for duplicates in ProposalList
    - remove metadata section for PUT and POST mapping as handled by ODA
+* STAR-635: 
+   - Correct Sensitivity Calculator Results for Confusion Noise, Total Sensitivity and Weighted Sensitivity
 * STAR-570:
    - Added Delete PDF button in Science and Technical Page
    - Upgrade to version 2.1.0 of ska-oso-pht-services